set http errorcode if an error occured; differend redis cachetimes for errorpages (#379)
* Add checks for all http-request parameters (days, abbreviation, ags),
therefor add three funktions, two constants and a new error class
to utils.ts. All checks are done directly after reciving the parameters.
If a parameter is not okay a error messege will be send with
http-status 400 (bad request)!
Changed the http-status for other erros to (eg. RKIErrors to 503
and axios errors to 503 too. all other errors are set to 400.
changed the redis caching time for errormessages too. Only succsessfull
 requests (http status 200) are cached for 1800 sec now.
I hope you like it.

Signed-off-by: Peter Markewitz <peter@markewitz.org>

* http-statuscode for RKI-Errors from errormessage, little changes in paramErrormessage

* remove parameter checks, not realy needed

* remove other changes from districts.ts

* remove other changes from germany.ts

* remove changes from states.ts

* remove changes from vaccination.ts

* adjust cache times

* little improvement in handling axios errors

* add check if days is a valid number in some funktions

* Add checks for :days against "null" AND "0"

Signed-off-by: Peter Markewitz <peter@markewitz.org>